- Moved GeneralReport to separate file
- Removed unnecessary getOut function
- Added methods sorting
