+++ src/org/traccar/reports/Events.java
@@ -30,7 +30,7 @@ public final class Events {
         return json.build().toString();
     }
 
-    public static byte[] getCsv(long userId, Collection<Long> deviceIds, Collection<Long> groupIds,
+    public static String getCsv(long userId, Collection<Long> deviceIds, Collection<Long> groupIds,
             Collection<String> types, Date from, Date to) throws SQLException {
         CsvBuilder csv = new CsvBuilder();
         csv.addHeaderLine(new Event());
